1) further aligned the base classes of FocalHead; 2) renamed roi_head -> img_roi_head; box_head -> pts_bbox_head

+++ contrib/petr/streampetr_head.py
@@ -745,7 +745,7 @@ class StreamPETRHead(AnchorFreeHead):
         # print(gt_bboxes.size(), bbox_pred.size())
         # DETR
         if sampling_result.num_gts > 0:
-            bbox_targets[pos_inds] = sampling_result.pos_gt_bboxes
+            bbox_targets[pos_inds] = sampling_result.pos_gt_bboxes.to(dtype=bbox_targets.dtype)
             bbox_weights[pos_inds] = 1.0
             labels[pos_inds] = gt_labels[sampling_result.pos_assigned_gt_inds]
         return (labels, label_weights, bbox_targets, bbox_weights,
